Improving focus visible per John Sweet a11y audit
From a11y audit: "When focus is on the menu items, the background color of the item under focus changes to gray. The contrast ratio of this gray color (rgb 66,66,66) to the original black color (rgb 34,34,34) is only 1.6:1—for low vision users, this may present some issues unless they know where to look. Make the background color for focused menu items a little bit brighter.  Also very difficult to perceive focus when it is on the [dropdown] ... subitems in the menu bar. Make focus on these items respond by changing background color."