Per #20513 a11y - switch underlines :hover :focus
Addressing Joe's a11y feedback: "The switch between the default view of links and the hover/focus view is slight; it would be very helpful you if added a change to that alternation that didn't depend on color. For :focus, you've got the outline, but for :hover, it's just a subtle color change. Removing underlines on :hover and :focus can be very effective." Removed default underlines and added them for h1 header.