Deleting space in "Theme Name" of "All Y'all"
To cope with a problem that the wordpress.org theme uploader has in creating directory names. It named this one's directory "all-yall" -- although it does not do this for native themes such as "Twenty Fifteen" which has a slug of "twentyfifteen" and directory of "twentyfifteen".  With the change the uploader should change it to "a11yall" so that it matches the slug.